requirements.txt
Flask
APScheduler
selenium
webdriver-manager
python-dotenv
gunicorn
